HiRes: tweak border relief
The {MWM,Fvwm}Border styles were written for 96 DPI displays.  But if
one has a 4K screen, the relief usually more prominent isn't seen so
well for higher resolution displays.

This visually hard-codes the areas neccessary for this change, and
should calculate this based on the DPI value of the server for better
scaling.